Add documentation for heading atoms
I'm not sure what the best practice is here, but it seems like several of the heading atoms could be collapsed into one pattern, with a more robust data model / variable structure.   

This commit documents all atoms as they are, and only deprecates the column heading as it isn't used anywhere.