Make callout links card a link
Originally, these cards were divs with children links
and JS would fire that open the link if someone clicked the card.
In html 5, we can have block links - meaning divs, h1, p can
be children of the 'a' element. This removes the need
for JS and will unblock this clash of JS for this element.